app/requirements.txt
Django==2.2.6
psycopg2-binary==2.8.3
gunicorn==19.9.0
djangorestframework==3.11
requests==2.23.0
numpy==1.18.3
django-cacheops==4.2
